requirements.txt
requests
beautifulsoup4
python-dotenv
langchain
langchain-core
langchain-community
langchain_groq
langgraph
langchain_tavily
tavily-python
Flask
Flask-SQLAlchemy
SQLAlchemy
supabase
typing_extensions
chromadb
gunicorn
gpt4all
